requirements.txt
absl-py==0.8.1
appdirs==1.4.3
astor==0.8.0
Click==7.0
colorama==0.4.1
coloredlogs==10.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
gekko==0.2.6
grpcio==1.25.0
h5py==2.10.0
humanfriendly==4.18
imageio==2.6.1
itk==5.0.1
itk-core==5.0.1
itk-filtering==5.0.1
itk-io==5.0.1
itk-numerics==5.0.1
itk-registration==5.0.1
itk-segmentation==5.0.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mock==3.0.5
natsort==6.2.0
networkx==2.4
numpy==1.16.4
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.0
psutil==5.6.7
pyopencl==2019.1.2
pyparsing==2.4.5
pyreadline==2.1
python-dateutil==2.8.1
pytools==2019.1.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.3
seaborn==0.9.0
SimpleITK==1.2.4
six==1.13.0
tabulate==0.8.6
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
voluptuous==0.11.7
vtk==8.1.2
Werkzeug==0.16.0